[CHERI-RISC-V] Provide a valid stack to the called function
Running the unit tests has exposed a few bounds violations due to using
the wrong stack pointer in the callee. Passing alloc_base results in a
capability bounded to the size of the alloca, whereas the expected
behaviour is to pass the remaining stack starting at the alloca (it can
be clobbered once the assembly code has loaded all the registers).