[Morello] Fix passing small structs in registers
When passing something like struct {long;long;} the existing code was
copying the raw struct bytes into the xregs array (if there is sufficient
space to pass the struct in memory). However, the xregs array holds
capabilities for Morello purecap, so this 16-byte memcpy means the struct
with two longs was stored in a single capability register.

We have to split up this copy into 8-byte chunks for each register to
avoid writing to the capability metadata instead of the xreg subset.
While fixing this problem the commit also adds logic to deal with structs
that contain capabilities, in which case we can pass up to 32 bytes in
registers instead of the aarch64 16-byte limit.