Use uintptr_t * when referencing a pointer value
The avalue vector of pointers passed to ffi_call() can include pointers
referencing pointer values. In such case, CHERI C requires that
a variable storing an avalue vector element must have a data type
indicating that it can reference a pointer value.

Use uintptr_t * for the avalue vector element to indicate the above
fact, including the source argument in extend_integer_type() that
derefences the avalue element.

libffi v3.4.4 replaced dereferences with memcpy(3) calls [1] in
extend_integer_type() that resulted in an underaligned load of
capability warning for ffi_call_int(). Such warning is a false negative
in this case and could be silenced by casting the source variable to
uintptr_t * when handling a pointer in extend_integer_type() but
updating the avalue element data type seems to be better a approach to
reflect the intentions here.

[1] https://github.com/libffi/libffi/commit/3874fd9865eef72a5cf5a761781b51f8db9fa267