Handle more than 8 arguments in CheriABI
Function calls with more than 8 arguments require passing 9th and later
arguments through memory referenced by the c9 register.

libffi has already passed such arguments on the stack. Extend the
implementation to store a pointer to the stack in c9 with bounds
restricted to the in-memory arguments. Additionally, use precise bounds
for all objects passed to ffi_call_SYSV(). They are allocated together
with the on-stack arguments and we want to make sure they do not overlap
with anything that should not be leaked to the callee.

Before we allocate memory for the on-stack arguments, we must first
compute the layout of data to be stored in memory that would meet
requirements of precise bounds to be used with that data. The number of
requested stack bytes is calculcated prior to ffi_call_int() and is
sufficient to hold on-stack function arguments and structure objects
passed by value. libffi allocates memory for a call context, on-stack
arguments, a call frame and return values on the caller's stack using
alloca(). In order to align the on-stack arguments to use precise
bounds, all sub-objects within that allocation must be appropriately
padded with respect to other sub-objects' alignments.  Otherwise, if
only the on-stack arguments were aligned, sub-objects placed after the
on-stack arguments could be misaligned.

After that, the stack can be further narrowed to use precise bounds only
for on-stack function arguments, excluding the structure objects passed
by value. Since the capability alignment grows with length and is
divisible by smaller alignments, it is safe to do that.

Co-authored-by:	Dapeng Gao <dapeng.gao@cl.cam.ac.uk>
Co-authored-by: Jessica Clarke <jessica.clarke@cl.cam.ac.uk>